requirements.txt
numpy==1.24.1
pandas==1.5.3
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
rsa==4.9
six==1.16.0
Telethon==1.26.1
